sensors/virtualsensors/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
cheroot==8.5.2
CherryPy==18.6.0
idna==2.10
influx-client==1.9.0
influxdb-client==1.17.0
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==3.3.0
jaraco.text==3.5.0
more-itertools==8.8.0
numpy==1.20.3
paho-mqtt==1.5.1
pandas==1.2.4
portend==2.7.1
python-dateutil==2.8.1
python-telegram-bot==13.6
pytool==3.16.1
pytz==2021.1
requests==2.25.1
Rx==3.2.0
simplejson==3.17.2
six==1.16.0
tempora==4.0.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.5
zc.lockfile==2.0
